refactor: centralize artifact extraction helpers in session-store.ts
Agent-Logs-Url: https://github.com/sabbour/kickstart/sessions/182c5e50-d150-4d5f-b8df-b72bd9598bc6
